NewUserFormDTO.java - Renamed method for naming consistency.
UpdatePasswordDTO.java
- Removed unused import.

UserAccountDetailsRowmapper.java
- Changed method for retrieving account ID to getLong to reflect target object field type.

UserDetailsServiceImpl.java
- Added method for loading all user accounts.
- Added method for loading a user account with a specified ID.
- Added method for loading a list of all teams.

+++ src/main/java/com/gaatvul/bugtracker/Rowmappers/UserAccountDetailsRowmapper.java
@@ -17,7 +17,7 @@ public class UserAccountDetailsRowmapper implements RowMapper<UserAccountDTO> {
         
         UserAccountDTO userAccountDetails = new UserAccountDTO();
 
-        userAccountDetails.setId(rs.getInt("account_id"));
+        userAccountDetails.setId(rs.getLong("account_id"));
         userAccountDetails.setFirstName(rs.getString("first_name"));
         userAccountDetails.setLastName(rs.getString("last_name"));
         userAccountDetails.setEmailAddress(rs.getString("email_address"));
